Highlight unreserved keywords
Fixes #27
